+++ jest/setupTests.ts
@@ -2,6 +2,5 @@ import enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 enzyme.configure({
-  adapter: new Adapter(),
-  disableLifecycleMethods: true
+  adapter: new Adapter()
 });
