+++ package.json
@@ -29,7 +29,7 @@
     "postbuild": "node tools/buildStylesheets.mjs",
     "test": "jest",
     "test:watch": "jest --watch",
-    "eslint": "eslint --ext js,mjs,ts,tsx --max-warnings 0 -f codeframe --cache --color examples src test tools",
+    "eslint": "eslint --ext js,mjs,ts,tsx --max-warnings 0 -f codeframe --cache --color examples src jest tools",
     "eslint:fix": "npm run eslint -- --fix",
     "typecheck": "tsc -p tsconfig.all.json",
     "prepublishOnly": "npm install && npm run build",
