+++ packages/react-data-grid-addons/src/data/__tests__/RowGrouper.test.js
