+++ src/HeaderCell.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import HeaderCell, { HeaderCellProps } from '../HeaderCell';
-import { ValueFormatter } from '../formatters';
-import { CalculatedColumn } from '../common/types';
+import HeaderCell, { HeaderCellProps } from './HeaderCell';
+import { ValueFormatter } from './formatters';
+import { CalculatedColumn } from './common/types';
 
 interface Row {
   bla?: string;
