+++ src/HeaderRow.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import helpers, { Row } from './GridPropHelpers';
-import HeaderRow, { HeaderRowProps } from '../HeaderRow';
-import HeaderCell from '../HeaderCell';
-import { DEFINE_SORT } from '../common/enums';
+import helpers, { Row } from './test/GridPropHelpers';
+import HeaderRow, { HeaderRowProps } from './HeaderRow';
+import HeaderCell from './HeaderCell';
+import { DEFINE_SORT } from './common/enums';
 
 describe('Header Row Unit Tests', () => {
   const defaultProps: HeaderRowProps<Row, 'id'> = {
