+++ src/editors/SimpleTextEditor.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import SimpleTextEditor from '../SimpleTextEditor';
-import { ValueFormatter } from '../../formatters';
-import { CalculatedColumn } from '../../common/types';
+import SimpleTextEditor from './SimpleTextEditor';
+import { ValueFormatter } from '../formatters';
+import { CalculatedColumn } from '../common/types';
 
 interface Row { text: string }
