+++ src/utils/columnUtils.test.ts
@@ -1,6 +1,6 @@
-import { canEdit } from '../columnUtils';
-import { ValueFormatter } from '../../formatters';
-import { CalculatedColumn, Omit } from '../../common/types';
+import { canEdit } from './columnUtils';
+import { ValueFormatter } from '../formatters';
+import { CalculatedColumn, Omit } from '../common/types';
 
 interface Row {
   PlacementType?: string;
