+++ src/utils/keyboardUtils.test.ts
@@ -1,4 +1,4 @@
-import { isCtrlKeyHeldDown, isKeyPrintable } from '../keyboardUtils';
+import { isCtrlKeyHeldDown, isKeyPrintable } from './keyboardUtils';
 
 describe('isKeyPrintable', () => {
   it('should return true if ctrl key is pressed', () => {
